Changes in PlayerMovement script
